+++ tests/unit/simulation/test_flight.py
@@ -413,6 +413,37 @@ def test_max_values(flight_calisto_robust):
     assert pytest.approx(285.94948, rel=rtol) == test.max_speed
 
 
+@pytest.mark.parametrize(
+    "flight_time_attr",
+    ["t_initial", "out_of_rail_time", "apogee_time", "t_final"],
+)
+def test_axial_acceleration(flight_calisto_custom_wind, flight_time_attr):
+    """Tests the axial_acceleration property by manually calculating the
+    dot product of the acceleration vector and the attitude vector at
+    specific time steps.
+
+    Parameters
+    ----------
+    flight_calisto_custom_wind : rocketpy.Flight
+        Flight object to be tested.
+    flight_time_attr : str
+        The name of the attribute of the flight object that contains the time
+        of the point to be tested.
+    """
+    flight = flight_calisto_custom_wind
+    t = getattr(flight, flight_time_attr)
+
+    calculated_axial_acc = flight.axial_acceleration(t)
+
+    expected_axial_acc = (
+        flight.ax(t) * flight.attitude_vector_x(t)
+        + flight.ay(t) * flight.attitude_vector_y(t)
+        + flight.az(t) * flight.attitude_vector_z(t)
+    )
+
+    assert pytest.approx(expected_axial_acc, abs=1e-9) == calculated_axial_acc
+
+
 def test_effective_rail_length(flight_calisto_robust, flight_calisto_nose_to_tail):
     """Tests the effective rail length of the flight simulation. The expected
     values are calculated by hand, and should be valid as long as the rail